All systems interact withe the DCS world plugin now.
Note that this interprets and displays positions in terms of world coordinates, not relative to the cockpit, which is incorrect. For example, if upside down and the target is to the left in world coordinates, the RWR should indicate target to the right. Currently, the system indicates it is still to the left, which is incorrect.